sc8r/requirements.txt
pyyaml>=6.0
requests>=2.28.0
click>=8.1.0
rich>=12.0.0
typer>=0.6.0
colorama>=0.4.5
cryptography>=38.0.0
pycryptodome>=3.15.0
tenacity>=8.0.0
jsonschema>=4.0.0
psutil>=5.0.0